Enable openstack-common messaging changes
As default messaging attributes have just been proposed to be added
to openstack-common, these changes rework openstack-image to take
advantage of that.  In these changes, I bump the openstack-common
dependency to 0.3.1 and enable rabbit in api.rb. This allows a user
to avoid searching for the rabbit node (but also provides the ability
for us to add search later) and is pluggable enough such that we can
later introduce alternative message queue services. As we are not
searching for the messaging nodes, I also remove those (now) unused
attributes.

Change-Id: Id462d0564d8a2fbdbab2ab7e6ce0953515dbb3ac